# docs/correspondence-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ehsan Correspondence API",
    "version": "0.3.0",
    "description": "Session-authenticated correspondence module. Cookie session and CSRF required for writes. External OAuth2/SSO/signature integrations are not configured."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "OAuth2": [
        "read"
      ]
    },
    {
      "sessionCookie": []
    }
  ],
  "paths": {
    "/api/communications/": {
      "get": {
        "summary": "List visible mail; kind/status/priority/secrecy/q filters",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid fields/state/reason"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "404": {
            "description": "Record not visible or missing"
          }
        }
      },
      "post": {
        "summary": "Create incoming or outgoing draft",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid fields/state/reason"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "404": {
            "description": "Record not visible or missing"
          }
        },
        "parameters": [
          {
            "name": "X-CSRFToken",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailInput"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      },
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/o/authorize/",
            "tokenUrl": "/o/token/",
            "scopes": {
              "read": "Read authorized data",
              "write": "Write authorized data"
            }
          }
        }
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      }
    },
    "schemas": {
      "MailInput": {
        "type": "object",
        "required": [
          "kind",
          "subject",
          "party",
          "classification",
          "unit",
          "owner"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "IN",
              "OUT",
              "INT",
              "CIR",
              "DEC",
              "RPL"
            ]
          },
          "subject": {
            "type": "string",
            "maxLength": 250
          },
          "body": {
            "type": "string"
          },
          "party": {
            "type": "integer"
          },
          "classification": {
            "type": "integer"
          },
          "unit": {
            "type": "integer"
          },
          "owner": {
            "type": "integer"
          },
          "external_number": {
            "type": "string",
            "description": "Required for IN"
          },
          "original_date": {
            "type": "string",
            "format": "date",
            "description": "Required for IN"
          },
          "priority": {
            "type": "string",
            "enum": [
              "normal",
              "important",
              "urgent",
              "critical"
            ]
          },
          "secrecy": {
            "type": "string",
            "enum": [
              "public",
              "internal",
              "restricted",
              "secret"
            ]
          },
          "signer": {
            "type": "integer"
          },
          "route": {
            "type": "integer"
          },
          "related": {
            "type": "integer",
            "description": "Required for RPL"
          },
          "no_attachments": {
            "type": "boolean"
          },
          "authorized": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "cc": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          }
        }
      },
...
    }
  }
}